requirements.txt
streamlit
pandas
geopandas
openai
langchain
scikit-learn
python-dotenv
matplotlib
langchain
azure-storage-blob
faiss-cpu
langchain-community
langchain-openai